fix: resolve axios API instance require cycle warning
